requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
pydantic==2.11.7
openai==1.99.9
azure-cognitiveservices-speech==1.45.0